Generate the summary of a remstimate object

A function that returns the summary of a remstimate object.

Usage

# S3 method for class 'remstimate'
summary(object, ...)

Arguments

object

is a remstimate object.

...

further arguments to be passed to the 'summary' method.

Value

no return value. Prints out the summary of a 'remstimate' object. The output can be save in a list, which contains the information printed out by the summary method.

Examples

# ------------------------------------ #
#       method 'summary' for the       #
#       tie-oriented model: "BSIR"     #
# ------------------------------------ #

# loading data
data(tie_data)

# processing event sequence with remify
tie_reh <- remify::remify(edgelist = tie_data$edgelist, model = "tie")
  
# specifying linear predictor
tie_model <- ~ 1 + 
               remstats::indegreeSender()+
               remstats::inertia()+
               remstats::reciprocity() 

# calculating statistics
tie_reh_stats <- remstats::remstats(reh = tie_reh, 
                                    tie_effects = tie_model)

# running estimation
tie_mle <- remstimate::remstimate(reh = tie_reh,
                                  stats = tie_reh_stats,
                                  method = "BSIR",
                                  nsim = 100,
                                  ncores = 1)

# summary
summary(tie_mle)
#> Relational Event Model (tie oriented) 
#> 
#> Call:
#> ~1 + remstats::indegreeSender() + remstats::inertia() + remstats::reciprocity()
#> 
#> 
#> Posterior Modes (BSIR with interval likelihood):
#> 
#>                Post.Mode   Post.SD     Q2.5%      Q50%  Q97.5% Pr(=0|x)
#> baseline       -4.930686  0.191566 -5.238398 -4.902682 -4.5320   <2e-16
#> indegreeSender  0.041427  0.034736 -0.021272  0.042923  0.1008   0.8308
#> inertia        -0.178813  0.086037 -0.355964 -0.213985 -0.0183   0.5357
#> reciprocity    -0.052704  0.097314 -0.246713 -0.054526  0.1148   0.8962
#> Log posterior: -605.362

# ------------------------------------ #
#      method 'summary' for the        #
#      actor-oriented model: "BSIR"    #
# ------------------------------------ #

# loading data
data(ao_data)

# processing event sequence with remify
ao_reh <- remify::remify(edgelist = ao_data$edgelist, model = "actor")
  
# specifying linear predictor (for sender rate and receiver choice model)
rate_model <- ~ 1 + remstats::indegreeSender()
choice_model <- ~ remstats::inertia() + remstats::reciprocity()

# calculating statistics
ao_reh_stats <- remstats::remstats(reh = ao_reh, 
                                   sender_effects = rate_model, 
                                   receiver_effects = choice_model)

# running estimation
ao_mle <- remstimate::remstimate(reh = ao_reh,
                                 stats = ao_reh_stats,
                                 method = "BSIR",
                                 nsim = 100,
                                 ncores = 1)

# summary
summary(ao_mle)
#> Relational Event Model (actor oriented) 
#> 
#> Call rate model **for sender**:
#> 
#> 	~1 + remstats::indegreeSender()
#> 
#> 
#> Posterior Modes rate model (BSIR with interval likelihood):
#> 
#>                 Post.Mode    Post.SD      Q2.5%       Q50%  Q97.5% Pr(=0|x)
#> baseline       -4.8318079  0.1945285 -5.1762069 -4.8465518 -4.4549   <2e-16
#> indegreeSender -0.0062247  0.0158862 -0.0360303 -0.0057925  0.0195   0.9025
#> Log posterior: -588.6842
#> -------------------------------------------------------------------------------- 
#> 
#> Call choice model **for receiver**:
#> 
#> 	~remstats::inertia() + remstats::reciprocity()
#> 
#> 
#> Posterior Modes choice model (BSIR with interval likelihood):
#> 
#>             Post.Mode   Post.SD     Q2.5%      Q50% Q97.5% Pr(=0|x)
#> inertia     -0.038097  0.082237 -0.174712 -0.038860 0.1146   0.8998
#> reciprocity  0.011089  0.069394 -0.130471  0.021295 0.1441   0.9080
#> Log posterior: -138.5375

# ------------------------------------ #
#   for more examples check vignettes  #
# ------------------------------------ #