Tie-Oriented Relational Event History

A randomly generated sequence of relational events with 5 actors and 100 events. The event sequence is generated by following a tie-oriented modeling approach (for more information run on console help(topic = remulateTie, package = "remulate") or ?remulate::remulateTie).

Usage

data(tie_data)

Format

tie_data is a list object containing the following objects:

edgelist

a data.frame with the raw simulated edgelist. The columns of the data.frame are:

time

the timestamp indicating the time at which each event occurred

actor1

the actor that generated the relational event

actor2

the actor that received the relational event

seed

the seed value used in remulate::remulateTie() for generating the event sequence

true.pars

a vector containing the values of the parameters used in the generation of the event sequence

Examples

# (1) load the data into the workspace
data(tie_data)

# (2) process event sequence with \code{remify}
tie_reh <- remify::remify(edgelist = tie_data$edgelist, model = "tie")

# (3) define linear predictor and claculate stastistcs with \code{remstats} package

## linear predictor
tie_model <- ~ 1 + remstats::indegreeSender() + remstats::inertia() + remstats::reciprocity() 

## calculate statistics
tie_reh_stats <- remstats::remstats(reh = tie_reh, tie_effects = tie_model)

# (4) estimate model using method = "MLE" and print out summary

## estimate model 
mle_tie <- remstimate::remstimate(reh = tie_reh, stats = tie_reh_stats, method = "MLE")

## print out a summary of the estimation
summary(mle_tie)
#> Relational Event Model (tie oriented) 
#> 
#> Call:
#> ~1 + remstats::indegreeSender() + remstats::inertia() + remstats::reciprocity()
#> 
#> 
#> Coefficients (MLE with interval likelihood):
#> 
#>                  Estimate   Std. Err    z value Pr(>|z|) Pr(=0)
#> baseline        -4.910454   0.187555 -26.181372   0.0000 <2e-16
#> indegreeSender   0.043490   0.036449   1.193170   0.2328 0.8307
#> inertia         -0.201506   0.088154  -2.285831   0.0223 0.4231
#> reciprocity     -0.052137   0.098237  -0.530728   0.5956 0.8968
#> Null deviance: 1216.739 on 100 degrees of freedom
#> Residual deviance: 1210.625 on 96 degrees of freedom
#> Chi-square: 6.11449 on 4 degrees of freedom, asymptotic p-value 0.1907597 
#> AIC: 1218.625 AICC: 1219.046 BIC: 1229.045