remstimate: Optimization Frameworks for Tie-Oriented and Actor-Oriented Relational Event Models

A comprehensive set of tools designed for optimizing likelihood within a tie-oriented (Butts, C., 2008, doi:10.1111/j.1467-9531.2008.00203.x ) or an actor-oriented modelling framework (Stadtfeld, C., & Block, P., 2017, doi:10.15195/v4.a14 ) in relational event networks. The package accommodates both frequentist and Bayesian approaches. The frequentist approaches that the package incorporates are the Maximum Likelihood Optimization (MLE) and the Gradient-based Optimization (GDADAMAX). The Bayesian methodologies included in the package are the Bayesian Sampling Importance Resampling (BSIR) and the Hamiltonian Monte Carlo (HMC). The flexibility of choosing between frequentist and Bayesian optimization approaches allows researchers to select the estimation approach which aligns the most with their analytical preferences.

See also

Author

Maintainer: Giuseppe Arena g.arena@uva.nl (ORCID)

Authors:

Rumana Lakdawala

Fabio Generoso Vieira

Other contributors:

Marlyne Meijerink-Bosman [contributor]

Diana Karimova [contributor]

Mahdi Shafiee Kamalabad [contributor]

Roger Leenders r.t.a.j.leenders@tilburguniversity.edu [contributor]

Joris Mulder j.mulder3@tilburguniversity.edu [contributor]