Generate the summary of a remstimate object

A function that returns the summary of a remstimate object.

Usage

# S3 method for class 'remstimate'
summary(object, ...)

Arguments

object

is a remstimate object.

...

further arguments to be passed to the 'summary' method.

Value

no return value. Prints out the summary of a 'remstimate' object. The output can be save in a list, which contains the information printed out by the summary method.

Examples

# ------------------------------------ #
#       method 'summary' for the       #
#       tie-oriented model: "BSIR"     #
# ------------------------------------ #

# loading data
data(tie_data)

# processing event sequence with remify
tie_reh <- remify::remify(edgelist = tie_data$edgelist, model = "tie")
  
# specifying linear predictor
tie_model <- ~ 1 + 
               remstats::indegreeSender()+
               remstats::inertia()+
               remstats::reciprocity() 

# calculating statistics
tie_reh_stats <- remstats::remstats(reh = tie_reh, 
                                    tie_effects = tie_model)

# running estimation
tie_mle <- remstimate::remstimate(reh = tie_reh,
                                  stats = tie_reh_stats,
                                  method = "BSIR",
                                  nsim = 100,
                                  ncores = 1)

# summary
summary(tie_mle)
#> Relational Event Model (tie oriented) 
#> 
#> Call:
#> ~1 + remstats::indegreeSender() + remstats::inertia() + remstats::reciprocity()
#> 
#> 
#> Posterior Modes (BSIR with interval likelihood):
#> 
#>                Post.Mode   Post.SD     Q2.5%      Q50%  Q97.5% Pr(=0|x)
#> baseline       -4.897778  0.183923 -5.316886 -4.936618 -4.5706   <2e-16
#> indegreeSender  0.052239  0.033067 -0.018439  0.049102  0.1094   0.7417
#> inertia        -0.209489  0.081257 -0.369031 -0.222512 -0.0413   0.2649
#> reciprocity    -0.065317  0.096046 -0.227090 -0.059380  0.1212   0.8881
#> Log posterior: -605.4761

# ------------------------------------ #
#      method 'summary' for the        #
#      actor-oriented model: "BSIR"    #
# ------------------------------------ #

# loading data
data(ao_data)

# processing event sequence with remify
ao_reh <- remify::remify(edgelist = ao_data$edgelist, model = "actor")
  
# specifying linear predictor (for sender rate and receiver choice model)
rate_model <- ~ 1 + remstats::indegreeSender()
choice_model <- ~ remstats::inertia() + remstats::reciprocity()

# calculating statistics
ao_reh_stats <- remstats::remstats(reh = ao_reh, 
                                   sender_effects = rate_model, 
                                   receiver_effects = choice_model)

# running estimation
ao_mle <- remstimate::remstimate(reh = ao_reh,
                                 stats = ao_reh_stats,
                                 method = "BSIR",
                                 nsim = 100,
                                 ncores = 1)

# summary
summary(ao_mle)
#> Relational Event Model (actor oriented) 
#> 
#> Call rate model **for sender**:
#> 
#> 	~1 + remstats::indegreeSender()
#> 
#> 
#> Posterior Modes rate model (BSIR with interval likelihood):
#> 
#>                 Post.Mode    Post.SD      Q2.5%       Q50%  Q97.5% Pr(=0|x)
#> baseline       -4.8033053  0.1810062 -5.1223591 -4.7889548 -4.4948   <2e-16
#> indegreeSender -0.0088236  0.0148676 -0.0426069 -0.0075654  0.0181   0.8934
#> Log posterior: -588.6746
#> -------------------------------------------------------------------------------- 
#> 
#> Call choice model **for receiver**:
#> 
#> 	~remstats::inertia() + remstats::reciprocity()
#> 
#> 
#> Posterior Modes choice model (BSIR with interval likelihood):
#> 
#>             Post.Mode   Post.SD     Q2.5%      Q50% Q97.5% Pr(=0|x)
#> inertia     -0.033359  0.074446 -0.191048 -0.039993 0.1168   0.9004
#> reciprocity  0.028577  0.071405 -0.102392  0.031216 0.1512   0.9023
#> Log posterior: -138.5382

# ------------------------------------ #
#   for more examples check vignettes  #
# ------------------------------------ #